Conflicts:
	app/controllers/application_controller.rb
	app/controllers/submitted_content_controller.rb
	app/controllers/users_controller.rb
	vendor/gems/hoptoad_notifier-2.4.4/MIT-LICENSE

+++ app/helpers/file_helper.rb
@@ -22,10 +22,14 @@ module FileHelper
   
   def self.update_file_location(oldpath,newpath)
     begin
-      create_directory_from_path(newpath)    
-      oldcontents = Dir.glob(oldpath + "/*")        
-      FileUtils.mv(oldcontents,newpath)
-      FileUtils.remove_dir(oldpath)
+      if oldpath and newpath
+        create_directory_from_path(newpath)    
+        oldcontents = Dir.glob(oldpath + "/*")        
+        FileUtils.mv(oldcontents,newpath)
+        FileUtils.remove_dir(oldpath)
+      elsif newpath # nil oldpath
+        create_directory_from_path(newpath)    
+      end
     rescue
       puts $!
     end
@@ -50,10 +54,10 @@ module FileHelper
     begin
       entries = Dir.entries(in_object.get_path)    
       if entries and entries.size == 2
-         FileUtils.remove_dir(in_object.get_path)          
-     end 
+        FileUtils.remove_dir(in_object.get_path)          
+      end 
     rescue PathError
-      # No action required 
+      # No action required
     rescue
       raise "An error was encountered while deleting the directory: "+$!
     end      
@@ -84,4 +88,4 @@ module FileHelper
       raise "An error was encountered while creating this directory: "+$!
     end   
   end   
-end
+end
